wip2

+++ backend/frogcrew/src/main/java/edu/tcu/cs/frogcrew/crewmember/CrewMemberController.java
@@ -28,7 +28,7 @@ public class CrewMemberController {
     }
 
     @GetMapping("/{crewId}")
-    public Result findCrewMemberById(@PathVariable Integer crewId) {
+    public Result findCrewMemberById(@PathVariable Long crewId) {
         CrewMember foundCrewMember = crewMemberService.findById(crewId);
         CrewMemberDto foundCrewMemberDto = crewMemberToCrewMemberDtoConverter.convert(foundCrewMember);
         return new Result(true, StatusCode.SUCCESS, "Find One Success", foundCrewMemberDto);
